feat: refactor EuroAssetAdapter to allow other fiat currencies
